\name{checkDuplicates_vect}
\alias{checkDuplicates_vect}
\title{
Check items that occur more than once
}
\description{
Check if a vector has any items that occur more than once and return a frequency table.
}
\usage{
checkDuplicates_vect(vect)
}
\arguments{
  \item{vect}{
Any types of vector in R
}
}
\author{
Peter I-Fan Wu
}
\examples{
checkDuplicates_vect(c(1,1,2,3,4,4,4,5,6,7,8,9,10))
}